update: move SBL to schemas
- creates consistency with the code structure
