+++ src/schemas/index.ts
@@ -3,6 +3,7 @@ export { brillSimple } from "./brillSimple.js";
 export { jss } from "./jss.js";
 export { michiganClaremont } from "./michiganClaremont.js";
 export { romaniote } from "./romaniote.js";
+export { SBL } from "./sblAcademic.js";
 export { sblAcademicSpirantization } from "./sblAcademicSpirantization.js";
 export { sblSimple } from "./sblSimple.js";
 export { tiberian } from "./tiberian.js";
